Remove redirects, replace home.tsx with index.tsx
